feat(authentication): support for google oauth2 authentication

* deprecated --name and --email parameters (settings the params is non blocking, a warning is displayed)
* updated README
* be explicit about scopes. At the moment, the profile scope is not required and the email is taken as a fallback on missing name
